+++ CHANGES.rst
@@ -64,7 +64,7 @@ Changelog
 
 - Added meta package for typing functionality.
 
-0.0.8 (????-??-??)
+0.0.8 (2019-09-19)
 -------------------
 
 - Fixed bug in get_argument_to_typevar.